requirements.txt
gym
gym-minigrid
pyro-ppl
pytorch